requirements.txt
Packages needed:
sys
os
string
numpy
sympy
time
matplotlib.pyplot